big_update
check_input:
-добавлен метод для проверки вводимого слова, может потом улучшу его чтобы можно было вводить предложение
-поменял в разных местах check_input на check_input_int чтобы было меньше приведений
pipe:
-добавлено поле имя для труб, так как по ним нужен поиск
-доработаны вводы и выводы в связи с новым параметром
station:
-эффективность теперь зависит от количества станций
main:
-введен новый функционал по поиску с фильтром(и редактированием по желанию)

+++ station.h
@@ -12,7 +12,7 @@ public:
     std::string name;
     int quantity;
     int quantity_in_work;
-    double efficiency;
+    int efficiency;
 
     int get_id() const;
